Revert back to using strings file
Lets try and keep these strings in the strings file so we don't start getting mixed strings all over the place.

I know it's a huge pain but I think it will come back to bite us in the long run (plus it will be a plus to tell him that some of the site is already translated to portugese and can be easily ported to other languages).